mac8021: Allow probe request injection on passive non radar
Based on section 11.1.4.3.8 (Non-scanning probe request transmission)
in Draft P802.11REVmd_D5.0 transmission of a probe request to an AP
known to be in range is allowed. Thus, allow probe request injection on
monitor interface in case the target AP is found in range and the
channel is not a radar channel.

[1] https://patchwork.kernel.org/project/linux-wireless/patch/iwlwifi.20210618133832.76e2f5ab0fe3.Iab8e73126dfe75a5448048ce118550bebda200b8@changeid/